moved cleaning rule to deid folder
